e2e/pyproject.toml
[project]
name = "e2e"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.57.0",
    "pytest-playwright>=0.7.2",
]
